aarch64: implement walking over the stack protector
Stepping into a function which contains the stack protector sequences
currently stops inside the prologue, resulting in us claiming to be on
the opening bracket or similar, instead of a useful statement inside the
function. Fix that by analysing the prologue instruction, and attempt to
walk through the sequence of instructions which set up the stack
protector canary.

gdb/
2014-06-03  Kyle McMartin  <kmcmarti@redhat.com>

	* aarch64-tdep.c (aarch64_skip_stack_chk_guard): New.
	(aarch64_skip_prologue): Skip over stack protector setup if possible.

gdb/testsuite/
2014-06-03  Kyle McMartin  <kmcmarti@redhat.com>

	* gdb.arch/aarch64-stack_chk_guard.c: New file.
	* gdb.arch/aarch64-stack_chk_guard.exp: New file.